backend/requirements.txt
aiomysql==0.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
backports.tarfile==1.2.0
bcrypt==5.0.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.9
dulwich==0.22.8
et_xmlfile==2.0.0
fastapi==0.109.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.6.0
msgpack==1.1.0
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pbs-installer==2025.3.17
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyMySQL==1.1.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.29
starlette==0.35.1
tomlkit==0.13.2
trove-classifiers==2025.3.19.19
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.27.0.post1
virtualenv==20.30.0
xattr==1.1.4
zipp==3.21.0
zstandard==0.23.0
